Add files via upload
Column Number of Fatalities, 2012

Mean = 92.34 , standard deviations = 92.01988, skewness = 2.821006, kurtosis =9.843226
median = 68.5

Since the mean and the standard deviation are almost the same so the difference between data is high and spread out which is to the right side. The distribution is not symmetrical. Since the skewness and kurtosis  are positive so it is positively skewed and leptokurtosis. For high standard deviation we see the outliers in the Boxplot for this column and very high value in the bar plot. It is looks right skewed and might be bimodal.


Column Penalties FY 2013 (Average$)

Mean =  1727.88, standard deviations = 963.3777,  skewness= 2.160074, kurtosis= 8.878504
Median= 1812
The graph shows the data  spreads out to the right which is not symmetrical. It is positively skewed and leptokurtosis. It is looks right skewed and might be bimodal (two peaks).
Since the standard deviation is high so we see the outlier in the boxplot and most of the data lies to the right side.


Column Number.of.Injuries.Illnesses.2012

Mean 64178.57, standard deviations =56633.63, skewness= 2.766831, kurtosis = 10.29762
Median= 62239.29
The mean and standard deviation are almost same so there is high difference between data. The data distribution is not symmetric and looks little bimodal. It has positive skewness and the box plot shows more data to the right. It is leptokurtosis. The boxplot shows most of the data lies to the right and the outliers of the data.


Column Number of inspector in each state 
Mean = 45.4, Standard Deviation = 40.91, Median = 30, Skewness = 1.85, Kurtosis = 10.29
The boxplot show the outliers of the data. We also can guess from the difference of number between mean and standard deviation which are the same. The Boxplot also show more data to the right. The skewness and kurtosis values shows it is right skewed and the histogram also shows bimodal and slope to the right. 




Column State Rank number 

Mean = 25.1, Med =25, SD= 14.37, skewness = 0.04, kurtosis = -1.29
This column is just the rank number. Even this column is numeric but it is just the rank number of each of the state. We have close values of mean, median and sd therefore the barplot shows no outliers. The skewness is to the right and platykurtic.